// src/managed/staging/authority-env.d.ts
/* eslint-disable */
// Generated by Wrangler by running `wrangler types src/managed/staging/authority-env.d.ts -c managed/staging/authority.json --env staging --strict-vars false --env-interface StagingAuthorityEnv --include-runtime false` (hash: c399c7602c2f8a9c2089ae700e89d053)
interface __BaseEnv_StagingAuthorityEnv {
  ENVIRONMENT: string;
  STAGING_ENABLED: string;
  STAGING_APPLICATION_ID: string;
  STAGING_INSTALLATION_ID: string;
  STAGING_CONTROL_HMAC_KEY: string;
  STAGING_AUTH_PEPPER: string;
  STAGING_AUTH_VERIFIER: string;
  STAGING_RECEIPT_HMAC_KEY: string;
  STAGING_SIGNING_KEYSET: string;
  STAGING_UNINSTALL_HMAC_KEY: string;
  STAGING_AUTHORIZATIONS: DurableObjectNamespace<import('./authority').StagingAuthorization>;
}
declare namespace Cloudflare {
  interface GlobalProps {
    mainModule: typeof import('./authority');
    durableNamespaces: 'StagingAuthorization';
  }
  interface Env extends __BaseEnv_StagingAuthorityEnv {}
}
interface StagingAuthorityEnv extends __BaseEnv_StagingAuthorityEnv {}
type StringifyValues<EnvType extends Record<string, unknown>> = {
  [Binding in keyof EnvType]: EnvType[Binding] extends string ? EnvType[Binding] : string;
};
declare namespace NodeJS {
  interface ProcessEnv extends StringifyValues<
    Pick<
      Cloudflare.Env,
      | 'ENVIRONMENT'
      | 'STAGING_ENABLED'
      | 'STAGING_APPLICATION_ID'
      | 'STAGING_INSTALLATION_ID'
      | 'STAGING_CONTROL_HMAC_KEY'
      | 'STAGING_AUTH_PEPPER'
      | 'STAGING_AUTH_VERIFIER'
      | 'STAGING_RECEIPT_HMAC_KEY'
      | 'STAGING_SIGNING_KEYSET'
      | 'STAGING_UNINSTALL_HMAC_KEY'
    >
  > {}
}

export { StagingAuthorityEnv };